Updated trackr_home.js
Hello! On the airplane on the way to Utah today and while in my Airbnb, I worked on this code. I added in a Stylesheet, new styles and a new look as well as got the code working for the button to increment the Opponent's life total.

That's all of the changes I made to this file!

Thank you!

.j.