trackr | Started the Expo Reconstruction
Hello GitHub! Tonight I started the process of recreating this app in Expo and started working on the layout and an error 404 "not found" file. Then it got late and I stopped working for the night. Going to continue to work on this and get it all re-built in Expo.

Thanks for reading!
.j.