Deleted trackr_home.js
This code now lives in "app/(tabs)/magic.js" so this code still lives on! ⭐ 
This is the *most important* of the old code so I couldn't get rid of it! It will be reworked mainlyinto the magic.js file and the Magic - Gameplay portion of the app, but also some of the basic life logic will make it's way into the Lorcana - Gameplay page too (lorcana.js).

Thank you for reading!
.j.