Updated magic.js
Decided tonight was the night to add in the old trackr code from two months ago into the magic.js file. I am going to work on making this look more like the storyboard vision idea I had for this page, but this was a good start!

The transform code on the your life buttons was being weird so I fixed that (wouldn't your life total be closest to you? anyways probably but it's late so goodnight lol)

Thanks for reading!
.j.