Update and renamed magic_LifeModal.js to MagicLifeModal.js
Hello GitHub! Today I went through the Modal tutorial on Expo and also scoured the internet for help and I added in code to the MagicLifeModal (formerly magic_LifeModal) that pops up a modal from the bottom. Next I'll add the logic to actually change the life totals.

It's a good start! Thanks for reading!
.j.