Updated MagicLifeModal.js
Worked on updating the color of the Modal and also the Max/Min Height/Width of it to make it look better. Also made the opacity look more like the Figma design. Need to fix the sizing of the Button text, want to make it more like the Figma design.

Thanks for reading!
.j.