Added MyContext.js
Created a file using the help of Google AI to create a context for the startingLifeTotal and get it passed into the Modal and through the Modal. I got that far and couldn't get the Magic.js to see the Context. Going to keep working on it! Thanks for reading!

.j.