Updated lorcana.js
Took the code from Magic.js and put it in here and updated the colors in line with the Figma design. Took out unnecessary code since once you hit 20 Lore the game is over. Need to figure out a way to stop the count from going up past that.

Thank you for reading!
.j.