Updated lorcana.js
Got a lot of good work on lorcana.js today! Made the styling like the updated styling for magic.js (which is closer to the Figma design), added logic in (with the help of Google :) ) that makes it so your lore can't go below 0 and above 20. Also changed "life" to "lore" to go with the game.

Lots of good changes
Thanks for reading!
.j.