Updated WelcomeModal.js
Worked on this over my lunch break and made it look good! Added in better styling, a paragraph that explains what the app is (straight from the README), and made the button be styled better.

Also changed the issue with the spacing realized I just had a flex: 1 in the container that I needed to remove.

I like how this looks now, excited to keep working on this! Thanks!

.j.