Created scrapped and scrapped lorcana.js
So this will be the big commit where I explain scrapped. /scrapped is where I am putting all of the code that will not be a part of the app but which I still want to be a part of the repository to show people what I have done!

I am shifting this app in a more Magic The Gathering-focused direction for right now and am looking to get a basic v1 of the app up and running, as well as a GitHub Page for the app as well as trying to at least get it online.

I did not feel that Lorcana functionality currently added to the project as a whole, so I am scrapping it for right now.

Feel free to look at what I did and use it for your own project!

Thank you and have a blessed night!

.j.