Updated magic.js
The next big change I wanted to make is make is so the user could add more players. Using ternary operators and the help of Google AI, we got the basic code in there.

I'm thinking I have it part of the life modal or make an entirely new modal that pops up after the life modal closes with an option to select how many players.

Goodnight GitHub!
.j.