Updated magic.js
Today was the day I added in the functionality of adding up to four players! This was a fun hour or so session of coding where I thought through the coding process and got chatgpt to help me work out the kinks!

Now the app works for you and your friends and we are getting close to v1!

Goodnight GitHub
.j.