Updated PlayerCountModal.js
Tonight this got working! Copied in the code from MagicLifeModal.js so I could quickly make it look the same and then got to work figuring out the buttons and where everything should go. Got help from ChatGPT on the context side of things multiple times but now it works and the feature is included!

.j.